Construct GB object
